requirements.txt
streamlit==1.36.0
mysql-connector-python==8.3.0
cryptography==42.0.7
opencv-python==4.9.0.80
face-recognition==1.3.0
numpy==1.26.4
pillow==10.3.0
insightface==0.7.3
pyotp==2.9.0
email-validator==2.2.0

